DBZ-8221 Add accessors to SchemaAndValueField
